Move Scheme back into Uri
