+++ client/src/main/scala/org/http4s/client/RequestKey.scala
@@ -1,8 +1,7 @@
 package org.http4s
 package client
 
-import org.http4s.Scheme
-import org.http4s.Uri.Authority
+import org.http4s.Uri.{Authority, Scheme}
 
 /** Represents a key for requests that can conceivably share a [[Connection]]. */
 final case class RequestKey(scheme: Scheme, authority: Authority)
