+++ core/src/main/scala/org/http4s/Uri.scala
@@ -1,14 +1,17 @@
 package org.http4s
 
+import cats._
 import cats.implicits.{catsSyntaxEither => _, _}
 import java.nio.charset.StandardCharsets
 import macrocompat.bundle
 import org.http4s.Uri._
-import org.http4s.internal.parboiled2.Parser
+import org.http4s.internal.parboiled2.{Parser => PbParser}
+import org.http4s.internal.parboiled2.CharPredicate.{Alpha, Digit}
 import org.http4s.parser._
 import org.http4s.syntax.string._
 import org.http4s.util._
 import scala.language.experimental.macros
+import scala.math.Ordered
 import scala.reflect.macros.whitebox.Context
 
 /** Representation of the [[Request]] URI
@@ -140,7 +143,7 @@ object Uri extends UriFunctions {
   /** Decodes the String to a [[Uri]] using the RFC 3986 uri decoding specification */
   def fromString(s: String): ParseResult[Uri] =
     new RequestUriParser(s, StandardCharsets.UTF_8).Uri
-      .run()(Parser.DeliveryScheme.Either)
+      .run()(PbParser.DeliveryScheme.Either)
       .leftMap(e => ParseFailure("Invalid URI", e.format(s)))
 
   /** Parses a String to a [[Uri]] according to RFC 3986.  If decoding
@@ -155,8 +158,65 @@ object Uri extends UriFunctions {
   /** Decodes the String to a [[Uri]] using the RFC 7230 section 5.3 uri decoding specification */
   def requestTarget(s: String): ParseResult[Uri] =
     new RequestUriParser(s, StandardCharsets.UTF_8).RequestUri
-      .run()(Parser.DeliveryScheme.Either)
+      .run()(PbParser.DeliveryScheme.Either)
       .leftMap(e => ParseFailure("Invalid request target", e.format(s)))
+
+  /** Each [[org.http4s.Uri]] begins with a scheme name that refers to a
+    * specification for assigning identifiers within that scheme.
+    *
+    * @see https://www.ietf.org/rfc/rfc3986.txt, Section 3.1
+    */
+  final class Scheme private (val value: String) extends Ordered[Scheme] {
+    override def equals(o: Any) = o match {
+      case that: Scheme => this.value.equalsIgnoreCase(that.value)
+      case _ => false
+    }
+
+    private[this] var hash = 0
+    override def hashCode(): Int = {
+      if (hash == 0) {
+        hash = hashLower(value)
+      }
+      hash
+    }
+
+    override def toString = s"Scheme($value)"
+
+    override def compare(other: Scheme): Int =
+      value.compareToIgnoreCase(other.value)
+  }
+
+  object Scheme {
+    val http: Scheme = new Scheme("http")
+    val https: Scheme = new Scheme("https")
+
+    def parse(s: String): ParseResult[Scheme] =
+      new Http4sParser[Scheme](s, "Invalid scheme") with Parser {
+        def main = scheme
+      }.parse
+
+    private[http4s] trait Parser { self: PbParser =>
+      def scheme = rule {
+        "https" ~ push(https) |
+          "http" ~ push(http) |
+          capture(Alpha ~ zeroOrMore(Alpha | Digit | "+" | "-" | ".")) ~> (new Scheme(_))
+      }
+    }
+
+    implicit val http4sInstancesForScheme: Show[Scheme] with HttpCodec[Scheme] with Order[Scheme] =
+      new Show[Scheme] with HttpCodec[Scheme] with Order[Scheme] {
+        def show(s: Scheme): String = s.toString
+
+        def parse(s: String): ParseResult[Scheme] =
+          Scheme.parse(s)
+
+        def render(writer: Writer, scheme: Scheme): writer.type =
+          writer << scheme.value
+
+        def compare(x: Scheme, y: Scheme) =
+          x.compare(y)
+      }
+  }
 
   type UserInfo = String
