+++ tests/src/test/scala/org/http4s/SchemeSpec.scala
@@ -2,6 +2,7 @@ package org.http4s
 
 import cats.implicits._
 import cats.kernel.laws.discipline.OrderTests
+import org.http4s.Uri.Scheme
 import org.http4s.internal.parboiled2.CharPredicate
 import org.http4s.testing.HttpCodecTests
 import org.http4s.util.Renderer
