+++ tests/src/test/scala/org/http4s/UriTemplateSpec.scala
@@ -1,6 +1,6 @@
 package org.http4s
 
-import org.http4s.Uri.{Authority, IPv4, IPv6, RegName}
+import org.http4s.Uri.{Authority, IPv4, IPv6, RegName, Scheme}
 import org.http4s.UriTemplate._
 
 object UriTemplateSpec extends Http4sSpec {
